Add an int64_t cast to allow float_emul.cc to build without warnings on NetBSD/pmax 5.0 with GCC.
git-svn-id: https://gxemul.svn.sourceforge.net/svnroot/gxemul/gxemul/trunk@5475 dbbdf1f2-f3a0-11dc-8807-f0e41a653009

+++ src/main/float_emul.cc
@@ -211,7 +211,7 @@ uint64_t ieee_store_float_value(double nf, int fmt, int nan)
 		 *  If nf < 0.0, then r2 will begin with a sequence of binary
 		 *  1's, which is ok.
 		 */
-		r3 = nf;
+		r3 = (int64_t) nf;
 		r2 = r3;
 		r |= r2;
